refactor: move type hint funcs to type hint utils
